card views look better

+++ app/src/main/java/com/example/android/cloudy/data/model/remote/DailyForecast.java
@@ -11,7 +11,7 @@ public class DailyForecast {
 
     @SerializedName("dt")
     @Expose
-    private Integer dt;
+    private Long dt;
     @SerializedName("temp")
     @Expose
     private Temp temp;
@@ -37,11 +37,11 @@ public class DailyForecast {
     @Expose
     private Double rain;
 
-    public Integer getDt() {
+    public Long getDt() {
         return dt;
     }
 
-    public void setDt(Integer dt) {
+    public void setDt(Long dt) {
         this.dt = dt;
     }
